Moving logging scripts to subfolder
