+++ Assets/Mirror/Editor/Logging/LogLevelWindow.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 353c7c9e14e82f349b1679112050b196
+guid: c3dbf48190d77d243b87962a82c3b164
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
