+++ Assets/Mirror/Runtime/Logging.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63d647500ca1bfa4a845bc1f4cff9dcc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
