+++ Assets/Mirror/Runtime/Logging/LogFactory.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 353c7c9e14e82f349b1679112050b196
+guid: d06522432d5a44e1587967a4731cd279
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
